% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster_positions}
\alias{cluster_positions}
\alias{cluster_helix}
\alias{cluster_chain}
\alias{cluster_satellite}
\title{Clusters of particles}
\usage{
cluster_positions(
  N = 5,
  cl_fun = cluster_chain,
  radius = 50,
  label = "Au",
  ...,
  out = "",
  digits = 8
)

cluster_helix(
  N = 5,
  a = 10,
  b = 10,
  c = 50,
  R0 = 100,
  pitch = 200,
  delta = pi/5,
  delta0 = 0,
  hand = 1,
  ...
)

cluster_chain(
  N = 5,
  pitch = 500,
  a = 50,
  b = 30,
  c = b,
  rot = rotation_euler_passive(0, 0, 0)
)

cluster_satellite(N = 30, Rcore = 30, Rsat = 4, gap = 0.1, exclusion = 10, ...)
}
\arguments{
\item{N}{number of particles}

\item{cl_fun}{cluster function}

\item{radius}{particle radius}

\item{label}{particle material label}

\item{...}{extra arguments passed to cl_fun}

\item{out}{filename}

\item{digits}{accuracy}

\item{a}{semi-axis}

\item{b}{semi-axis}

\item{c}{semi-axis}

\item{R0}{helix radius}

\item{pitch}{chain pitch}

\item{delta}{helix angle step}

\item{delta0}{helix start angle}

\item{hand}{helix handedness}

\item{rot}{rotation matrix applied to each particle}

\item{Rcore}{core radius}

\item{Rsat}{satellite radius}

\item{gap}{gap distance}

\item{exclusion}{minimum exclusion distance for hc positions}
}
\value{
returns scatterers positions and sizes for an input file
}
\description{
Defines various cluster geometries and exports in a format suitable for TERMS

Helix of particles

Chain of particles

Core-satellite cluster of spheres
}
\section{Functions}{
\itemize{
\item \code{cluster_positions()}: write cluster positions to input file

\item \code{cluster_helix()}: helical cluster

\item \code{cluster_chain()}: linear chain cluster

\item \code{cluster_satellite()}: core-satellite cluster

}}
\examples{
cluster_positions()
cluster_helix()
cluster_chain()
cluster_satellite()
}
