% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{sample_fibonacci}
\alias{sample_fibonacci}
\title{sample_fibonacci}
\usage{
sample_fibonacci(N = 301)
}
\arguments{
\item{N}{number of points}
}
\description{
Fibonacci coverage of a sphere
}
\details{
Produces a set of points that covers rather uniformly the unit sphere with N points
with a spiral-like pattern based on a Fibonacci sequence
}
\concept{low_level sample fibonacci sampling of a sphere}
