% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{consolidate_xsec}
\alias{consolidate_xsec}
\alias{consolidate_partials}
\title{Reshape cross-section results into a convenient format for post-processing and plotting}
\usage{
consolidate_xsec(hdf5, verbose = TRUE, ...)

consolidate_partials(hdf5, verbose = TRUE)
}
\arguments{
\item{hdf5}{filename}

\item{verbose}{logical: print attributes}

\item{...}{extra arguments passed to the final list}
}
\value{
returns a list containing data.frames in long format
}
\description{
Extracts commonly-used information from a HDF5 file storing far-field cross-sections (Mode=2), and reshapes the data into long-format data.frames suitable for plotting

Extracts partial absorption cross-sections from a HDF5 file storing far-field cross-sections (Mode=2), and reshapes the data into long-format data.frames suitable for plotting
}
\section{Functions}{
\itemize{
\item \code{consolidate_xsec}: consolidate cross-sections

\item \code{consolidate_partials}: consolidate partial absorption cross-sections for multilayered spheres
}}

