% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R, R/tmat.R, R/utils.R, R/visualise.R
\name{equal_sizes}
\alias{equal_sizes}
\alias{equal_angles}
\alias{p_index}
\alias{indices}
\alias{unpack_indices}
\alias{read_amat}
\alias{amat_to_tmatlist}
\alias{read_tmat}
\alias{display_amat}
\alias{display_tmat}
\alias{display_prestaged}
\alias{export_cubature}
\alias{rotation_euler_passive}
\alias{rotation_euler_active}
\alias{euler_to_axisangle}
\alias{symmetric_range}
\alias{get_geometry}
\alias{cluster_geometry}
\alias{visualise_rgl}
\title{Utility functions}
\usage{
equal_sizes(a, b, c, N)

equal_angles(phi, theta, gamma, N)

p_index(in1, in2)

indices(n_max = 3, n_part = 1)

unpack_indices(n_max = 3, j_max = 2)

read_amat(f)

amat_to_tmatlist(a, n_max = 3, n_part = 2)

read_tmat(f, save = FALSE)

display_amat(l)

display_tmat(s)

display_prestaged(a, n_max = 3, n_part = 2, draw = TRUE)

export_cubature(q = cubs::cubs(N = 10, cubature = "lebedev"), out = "")

rotation_euler_passive(phi, theta, psi)

rotation_euler_active(phi, theta, psi)

euler_to_axisangle(a, b, c)

symmetric_range(range)

get_geometry(input = "input")

cluster_geometry(ge)

visualise_rgl(cl, outfile = NULL, show_core = FALSE, ...)
}
\arguments{
\item{a}{Euler angle}

\item{b}{Euler angle}

\item{c}{Euler angle}

\item{N}{number of particles}

\item{phi}{Euler angle}

\item{theta}{Euler angle}

\item{gamma}{Euler angle}

\item{in1}{index}

\item{in2}{index}

\item{n_max}{maximum order}

\item{n_part}{number of particles}

\item{j_max}{size of collective T-matrix}

\item{f}{filename}

\item{save}{store result as Rds file}

\item{l}{list of T-matrices, from amat_to_tmatlist}

\item{s}{T-matrices, from read_tmat}

\item{draw}{logical, draw output}

\item{q}{data.frame with angles and weights, from cubs::cubs()}

\item{out}{filename}

\item{psi}{Euler angle}

\item{range}{range (2-vector)}

\item{input}{filename}

\item{ge}{geometry, from get_geometry}

\item{cl}{'cluster' object}

\item{outfile}{optional output snapshot}

\item{show_core}{display a core sphere if 'R0' field present}

\item{...}{additional parameters passed to rgl.ellipsoids}
}
\description{
Matrix of equal sizes

Matrix of equal angles

Combined T-matrix index

T-matrix indices

Unpack T-matrix indices

Unpack T-matrix indices

Wrap staged matrices into a list of T-matrix like objects

Read T-matrix into long-format data.frame

Display staged matrix

Display T-matrix

Display prestaged matrix

Generate an incidence file for spherical cubature

Euler rotation matrix

Euler rotation matrix

Axis-angle rotation from Euler angles

Extend a range symmetrically about 0

Extract geometry information from input file

Wrap geometry information into a 'cluster' structure

Visualise a 'cluster' structure
}
\section{Functions}{
\itemize{
\item \code{equal_sizes}: equal sizes

\item \code{equal_angles}: equal angles

\item \code{p_index}: p-index

\item \code{indices}: indices

\item \code{unpack_indices}: unpack indices

\item \code{read_amat}: unpack indices

\item \code{amat_to_tmatlist}: wrap staged matrices

\item \code{read_tmat}: read T-matrix

\item \code{display_amat}: display staged matrix

\item \code{display_tmat}: display T-matrix

\item \code{display_prestaged}: display prestaged matrix

\item \code{export_cubature}: export a spherical cubature

\item \code{rotation_euler_passive}: passive rotation matrix

\item \code{rotation_euler_active}: active rotation matrix

\item \code{euler_to_axisangle}: axis-angle rotation

\item \code{symmetric_range}: symmetric range

\item \code{get_geometry}: extract geometry information from input file

\item \code{cluster_geometry}: wrap geometry information obtained from input file

\item \code{visualise_rgl}: rgl visualisation of a cluster
}}

