% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{extract_time}
\alias{extract_time}
\title{Extract timings from log files}
\usage{
extract_time(log = "log")
}
\arguments{
\item{log}{filename}
}
\value{
returns a tibble of timings
}
\description{
Parses a log file to extract timing information from subroutines (verbosity-dependent)
}
\section{Functions}{
\itemize{
\item \code{extract_time()}: extract timings from log files

}}
